Moved interfaces to new subdirectory
