+++ DeathmicChatbot/Counter.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using DeathmicChatbot.Interfaces;
 
 #endregion
