+++ DeathmicChatbot/DeathmicChatbot.csproj
@@ -62,12 +62,12 @@
     <Compile Include="Counter.cs" />
     <Compile Include="CounterEventArgs.cs" />
     <Compile Include="CountItem.cs" />
-    <Compile Include="ICounter.cs" />
-    <Compile Include="ITextFile.cs" />
+    <Compile Include="Interfaces\ICounter.cs" />
+    <Compile Include="Interfaces\ITextFile.cs" />
     <Compile Include="StreamInfo\Hitbox\HitboxProvider.cs" />
     <Compile Include="StreamInfo\Hitbox\HitboxStreamData.cs" />
     <Compile Include="TextFile.cs" />
-    <Compile Include="StreamInfo\IStreamProvider.cs" />
+    <Compile Include="Interfaces\IStreamProvider.cs" />
     <Compile Include="JoinLogger.cs" />
     <Compile Include="LogManager.cs" />
     <Compile Include="MessageQueue.cs" />
@@ -80,8 +80,8 @@
     </Compile>
     <Compile Include="Stream.cs" />
     <Compile Include="StreamData.cs" />
-    <Compile Include="ILogManagerProvider.cs" />
-    <Compile Include="IRestClientProvider.cs" />
+    <Compile Include="Interfaces\ILogManagerProvider.cs" />
+    <Compile Include="Interfaces\IRestClientProvider.cs" />
     <Compile Include="StreamInfo\Hitbox\Livestream.cs" />
     <Compile Include="LogManagerProvider.cs" />
     <Compile Include="StreamInfo\Hitbox\Request.cs" />
