+++ DeathmicChatbot/Interfaces/ILogManagerProvider.cs
@@ -5,7 +5,7 @@ using System.Diagnostics;
 #endregion
 
 
-namespace DeathmicChatbot
+namespace DeathmicChatbot.Interfaces
 {
     public interface ILogManagerProvider
     {
