+++ DeathmicChatbot/LogManagerProvider.cs
@@ -1,6 +1,7 @@
 ﻿#region Using
 
 using System.Diagnostics;
+using DeathmicChatbot.Interfaces;
 
 #endregion
