+++ DeathmicChatbot/Program.cs
@@ -11,6 +11,7 @@ using System.Net.Sockets;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
+using DeathmicChatbot.Interfaces;
 using DeathmicChatbot.Properties;
 using DeathmicChatbot.StreamInfo.Hitbox;
 using DeathmicChatbot.StreamInfo.Twitch;
